Estimate body surface area

Estimate body surface area (BSA)

bsa_adult(height_m, weight_kg, ...)

bsa_dubois_dubois(height_m, weight_kg, ...)

bsa_mosteller(height_m, weight_kg, ...)

bsa_haycock(height_m, weight_kg, ...)

bsa_gehan_george(height_m, weight_kg, ...)

bsa_boyd(height_m, weight_kg, ...)

bsa_fujimoto(height_m, weight_kg, ...)

bsa_takahira(height_m, weight_kg, ...)

bsa_shuter_aslani(height_m, weight_kg, ...)

bsa_schlich(height_m, weight_kg, male, ...)

Arguments

height_m
height(s) in meters

weight_kg
numeric vector of weight(s) in kg

...
passed to validation

male
logical value(s) whether patient is male. TRUE or FALSE.

Value

numeric vector of body surface areas in m^2.

Details

1.73 m^2 is commonly used as an average adult BSA.

Functions

bsa_adult: Uses `bsa_mosteller()`

bsa_dubois_dubois: Du Bois and Du Bois formula (usually the preferred formula for adults and children)

bsa_mosteller: Mosteller formula

bsa_haycock: Haycock formula

bsa_gehan_george: Gehan and George formula

bsa_boyd: Boyd formula

bsa_fujimoto: Fujimoto formula (often used for Japanese individuals)

bsa_takahira: Takahira formula (a variant of Du Bois that could be used for Japanese individuals; Fujimoto may be preferred)

bsa_shuter_aslani: Shuter and Aslani formula

bsa_schlich: Schlich formula

References

Du Bois D, Du Bois EF (Jun 1916). "A formula to estimate the approximate surface area if height and weight be known". Archives of Internal Medicine. 17 (6): 863–71.

Mosteller, RD (1987). "Simplified calculation of body-surface area". N Engl J Med. 317 (17): 1098. https://www.ncbi.nlm.nih.gov/pubmed/3657876

Haycock, GB, Schwartz, GJ, Wisotsky, DH (1978). Geometric method for measuring body surface area: A height-weight formula validated in infants, children and adults. J Pediatr. 93: 62–66.

Gehan EA, George SL, Cancer Chemother Rep 1970, 54:225-235

Boyd, Edith (1935). The Growth of the Surface Area of the Human Body. University of Minnesota. The Institute of Child Welfare, Monograph Series, No. x. London: Oxford University Press.

Fujimoto S, Watanabe T, Sakamoto A, Yukawa K, Morimoto K. Studies on the physical surface area of Japanese. 18. Calculation formulae in three stages over all ages. Nippon Eiseigaku Zasshi 1968;5:443–50.

Fujimoto S, Watanabe T, Sakamoto A, Yukawa K, Morimoto K. Studies on the physical surface area of Japanese. 18. Calculation formulae in three stages over all ages. Nippon Eiseigaku Zasshi 1968;5:443–50.

Shuter, B; Aslani, A (2000). "Body surface area: Du Bois and Du Bois revisited". European Journal of Applied Physiology. 82 (3): 250–254.

Schlich, E; Schumm, M; Schlich, M (2010). "3-D-Body-Scan als anthropometrisches Verfahren zur Bestimmung der spezifischen Körperoberfläche". Ernährungs Umschau. 57: 178–183.

Examples

bsa_dubois_dubois(2, 80)

#> [1] 2.154896

bsa_dubois_dubois(1.5, 80)

#> [1] 1.749225

stopifnot(
  identical(
    bsa_mosteller(1.5, 80),
    bsa_adult(1.5, 80)))