Get mean atmospheric pressure at given altitude in kPa

Get mean atmospheric pressure at given altitude in kPa

pres_atm_kPa(altitude_m)

pres_atm_frac(altitude_m)

Arguments

altitude_m
Altitude above mean sea level in meters

Value

Pressure in pascals

Functions

pres_atm_frac: Get fraction of mean atomspheric pressure at sea level

Examples

pres_atm_kPa(-430.5) # Dead Sea

#> [1] 106605

pres_atm_kPa(0)

#> [1] 101325

pres_atm_kPa(3440) # Namche Bazaar

#> [1] 66289.45

pres_atm_kPa(4260) # Dingboche

#> [1] 59602.43

pres_atm_kPa(5364) # Everest Base Camp

#> [1] 51474.86

pres_atm_kPa(6000) # Camp 1

#> [1] 47217.62

pres_atm_kPa(6400) # Camp 2

#> [1] 44689.9

pres_atm_kPa(7200) # Camp 3

#> [1] 39963.68

pres_atm_kPa(7950) # Camp 4

#> [1] 35909.55

pres_atm_kPa(8850) # Everest summit

#> [1] 31492

pres_atm_frac(8850) # fraction of sea level pressure on Everest

#> [1] 0.3108019